taximovil/requirements.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
certifi==2018.1.18
chardet==3.0.4
conekta==2.4.0
coreapi==2.3.3
coreschema==0.0.4
Django==2.0.3
django-datatables-view==1.16.0
django-filter==1.1.0
django-rest-framework==0.1.0
django-widgets-improved==1.5.0
djangorestframework==3.7.7
djangorestframework-gis==0.13
drf-yasg==1.4.7
drfdocs==0.0.11
fcm-django==0.2.19pi
future==0.16.0
googlemaps==3.0.2
idna==2.6
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10
kombu==4.1.0
Markdown==2.6.11
MarkupSafe==1.0
nose==1.3.7
openapi-codec==1.3.2
psycopg2==2.7.4
psycopg2-binary==2.7.4
pyfcm==1.4.5
Pygments==2.2.0
PyJWT==1.6.0
PySocks==1.6.8
pytz==2018.3
requests==2.18.4
requests-toolbelt==0.8.0
ruamel.yaml==0.15.35
simplejson==3.13.2
six==1.11.0
twilio==6.10.4
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4
